# swagger-spec.json
{
  "openapi": "3.0.0",
  "paths": {
...
    "/api/blog/posts": {
      "get": {
        "operationId": "BlogController_getAllPosts",
        "parameters": [
          {
            "name": "page",
            "required": false,
            "in": "query",
            "schema": {
              "type": "number"
            }
          },
          {
            "name": "limit",
            "required": false,
            "in": "query",
            "schema": {
              "type": "number"
            }
          },
          {
            "name": "category",
            "required": false,
            "in": "query",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "tag",
            "required": false,
            "in": "query",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": ""
          }
        },
        "summary": "Obtenir tous les articles publiés",
        "tags": [
          "Blog"
        ]
      },
      "post": {
        "operationId": "BlogController_createPost",
        "parameters": [],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/CreatePostDto"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": ""
          }
        },
        "security": [
          {
            "JWT-auth": []
          }
        ],
        "summary": "Créer un article",
        "tags": [
          "Blog"
        ]
      }
    },
    "/api/blog/posts/popular": {
      "get": {
        "operationId": "BlogController_getPopularPosts",
        "parameters": [
          {
            "name": "limit",
            "required": true,
            "in": "query",
            "schema": {
              "type": "number"
            }
          }
        ],
        "responses": {
          "200": {
            "description": ""
          }
        },
        "summary": "Obtenir les articles populaires",
        "tags": [
          "Blog"
        ]
      }
    },
    "/api/blog/posts/search": {
      "get": {
        "operationId": "BlogController_searchPosts",
        "parameters": [
          {
            "name": "q",
            "required": true,
            "in": "query",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": ""
          }
        },
        "summary": "Rechercher des articles",
        "tags": [
          "Blog"
        ]
      }
    },
    "/api/blog/posts/category/{category}": {
      "get": {
        "operationId": "BlogController_getPostsByCategory",
        "parameters": [
          {
            "name": "category",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": ""
          }
        },
        "summary": "Obtenir les articles par catégorie",
        "tags": [
          "Blog"
        ]
      }
    },
    "/api/blog/posts/{slug}": {
      "get": {
        "operationId": "BlogController_getPostBySlug",
        "parameters": [
          {
            "name": "slug",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "userId",
            "required": true,
            "in": "query",
            "schema": {
              "type": "number"
            }
          }
        ],
        "responses": {
          "200": {
            "description": ""
          }
        },
        "summary": "Obtenir un article par slug",
        "tags": [
          "Blog"
        ]
      }
    },
    "/api/blog/my/posts": {
      "get": {
        "operationId": "BlogController_getMyPosts",
        "parameters": [],
        "responses": {
          "200": {
            "description": ""
          }
        },
        "security": [
          {
            "JWT-auth": []
          }
        ],
        "summary": "Obtenir mes articles",
        "tags": [
          "Blog"
        ]
      }
    },
    "/api/blog/posts/{id}": {
      "put": {
        "operationId": "BlogController_updatePost",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "number"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/UpdatePostDto"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": ""
          }
        },
        "security": [
          {
            "JWT-auth": []
          }
        ],
        "summary": "Mettre à jour un article",
        "tags": [
          "Blog"
        ]
      },
      "delete": {
        "operationId": "BlogController_deletePost",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "number"
            }
          }
        ],
        "responses": {
          "200": {
            "description": ""
          }
        },
        "security": [
          {
            "JWT-auth": []
          }
        ],
        "summary": "Supprimer un article",
        "tags": [
          "Blog"
        ]
      }
    },
    "/api/blog/posts/{id}/publish": {
      "post": {
        "operationId": "BlogController_publishPost",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "number"
            }
          }
        ],
        "responses": {
          "201": {
            "description": ""
          }
        },
        "security": [
          {
            "JWT-auth": []
          }
        ],
        "summary": "Publier un article",
        "tags": [
          "Blog"
        ]
      }
    },
    "/api/blog/posts/{postId}/comments": {
      "post": {
        "operationId": "BlogController_addComment",
        "parameters": [
          {
            "name": "postId",
            "required": true,
            "in": "path",
            "schema": {
              "type": "number"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/CreateCommentDto"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": ""
          }
        },
        "security": [
          {
            "JWT-auth": []
          }
        ],
        "summary": "Ajouter un commentaire",
        "tags": [
          "Blog"
        ]
      }
    },
    "/api/blog/comments/{id}": {
      "put": {
        "operationId": "BlogController_updateComment",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "number"
            }
          }
        ],
        "responses": {
          "200": {
            "description": ""
          }
        },
        "security": [
          {
            "JWT-auth": []
          }
        ],
        "summary": "Modifier un commentaire",
        "tags": [
          "Blog"
        ]
      },
      "delete": {
        "operationId": "BlogController_deleteComment",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "number"
            }
          }
        ],
        "responses": {
          "200": {
            "description": ""
          }
        },
        "security": [
          {
            "JWT-auth": []
          }
        ],
        "summary": "Supprimer un commentaire",
        "tags": [
          "Blog"
        ]
      }
    },
    "/api/blog/posts/{postId}/like": {
      "post": {
        "operationId": "BlogController_toggleLike",
        "parameters": [
          {
            "name": "postId",
            "required": true,
            "in": "path",
            "schema": {
              "type": "number"
            }
          }
        ],
        "responses": {
          "201": {
            "description": ""
          }
        },
        "security": [
          {
            "JWT-auth": []
          }
        ],
        "summary": "Liker/Unliker un article",
        "tags": [
          "Blog"
        ]
      }
    },
...
  },
  "info": {
    "title": "EduLearn API",
    "description": "API pour la plateforme EduLearn",
    "version": "1.0",
    "contact": {}
  },
...
  "servers": [],
  "components": {
    "securitySchemes": {
      "JWT-auth": {
        "scheme": "bearer",
        "bearerFormat": "JWT",
        "type": "http"
      }
    },
    "schemas": {
...
      "CreatePostDto": {
        "type": "object",
        "properties": {
          "title": {
            "type": "string",
            "example": "Introduction à TypeScript"
          },
          "excerpt": {
            "type": "string",
            "example": "Un aperçu rapide de TypeScript"
          },
          "content": {
            "type": "string",
            "example": "Contenu complet du post..."
          },
          "featuredImage": {
            "type": "string"
          },
          "category": {
            "type": "string",
            "enum": [
              "technology",
              "design",
              "business",
              "marketing",
              "programming",
              "tutorial",
              "news",
              "other"
            ],
            "example": "programming"
          },
          "tags": {
            "example": [
              "typescript",
              "javascript"
            ],
            "type": "array",
            "items": {
              "type": "string"
            }
          },
          "commentsEnabled": {
            "type": "boolean",
            "example": true
          }
        },
        "required": [
          "title",
          "excerpt",
          "content",
          "category",
          "commentsEnabled"
        ]
      },
      "UpdatePostDto": {
        "type": "object",
        "properties": {}
      },
      "CreateCommentDto": {
        "type": "object",
        "properties": {
          "content": {
            "type": "string",
            "example": "Super article !"
          },
          "parentId": {
            "type": "number",
            "example": 1
          }
        },
        "required": [
          "content"
        ]
      },
...
    }
  }
}